+++ Sample1/sample1.cs
@@ -21,7 +21,7 @@ delegate void DateTimeChange(DateTime_t[] now);
 [FileName("sample1.cs")]
 interface Inter
 {
-    Int32 Add(Byte a, Byte b);
+    Int32 Add(Int32 a, Int32 b);
 
     bool NoArg();
     void NoReturn();
